New Transforms object + options + function changes
Implementation of the transformations object in the config YAML, Which contains the nested objects in a format like SOURCE->Destination.
There is a priority level nested within the SOURCE->Destination object, which is used in the engine queue, along with a confidence level and an exclude option.
The keywords of all and none can be used in the destination portion.
"All" would have to be determined from the engine perspective, while for "none", If there are multiple entries with the same source that contain a "none" destination at all, the config package will render it as broken, returning an error to the user.
Other options that are added into the config package is the engine URI for its api use and the confidence level globally for transforms.

There are function parameter changes that were removed because it was unused.

Dependencies were updated.